+++ sources/Graph/node.h
@@ -186,9 +186,9 @@ inline void
 Node::AddEdgeInDir( Edge *edge, GraphDir dir)
 {
     assert( isNotNullP( edge));
-    GRAPH_ASSERTD( GRAPH_DIR_DOWN == EDGE_LIST_SUCCS,
+    GRAPH_ASSERTD( (int) GRAPH_DIR_DOWN == (int) EDGE_LIST_SUCCS,
                    "Enums of direction and edge lists are not having right values");
-    GRAPH_ASSERTD( GRAPH_DIR_UP == EDGE_LIST_PREDS,
+    GRAPH_ASSERTD( (int) GRAPH_DIR_UP == (int) EDGE_LIST_PREDS,
                    "Enums of direction and edge lists are not having right values");
     edge->attach( dir, first_edge[ dir]); 
     first_edge[ dir] = edge;
